+++ CHANGELOG.md
@@ -6,6 +6,9 @@ directly impact users rather than highlighting other crucial architectural updat
 
 ## [Unreleased]
 
+### Added
+- Option to export SDK and wallet logs in `Export private data` screen.
+
 ### Changed
 - The background of Onboarding and some of the Settings screen has been updated to show a subtle texture of a grid pattern.
 - The sapling address + QR code has been restored on the Receive tab (for testnet only.)